requirements.txt
openai
langchain
tiktoken
pinecone-client
python-dotenv